pyproject.toml
[project]
name = "memory-core"
version = "0.1.0"
description = ""
authors = [
    {name = "Melody Lai"}
]
readme = "README.md"
requires-python = "3.12.5"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "llama-index (>=0.14.0,<0.15.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "llama-index-embeddings-langchain (>=0.4.1,<0.5.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "openai (>=1.107.0,<2.0.0)",
    "opencc (>=1.1.9,<2.0.0)",
    "pandas (<2.3.0)",
    "llvmlite (>=0.42,<0.45)",
    "descript-audio-codec (>=1.0.0,<2.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "transformers (>=4.45.1,<4.47.0)",
    "librosa (>=0.11.0,<0.12.0)",
    "dacite (>=1.9.2,<2.0.0)",
    "boto3 (==1.35.36)",
    "s3fs (>=2025.9.0,<2026.0.0)",
    "torchvision (>=0.23.0,<0.24.0)",
    "torchaudio (>=2.8.0,<3.0.0)",
    "json-repair (>=0.50.1,<0.51.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "vector-quantize-pytorch (>=1.23.2,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "ruff (==0.12.2)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "langid (>=1.1.6,<2.0.0)",
    "jieba (>=0.42.1,<0.43.0)",
    "accelerate (>=0.26.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pedalboard (>=0.9.17,<0.10.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
